+++ contracts/interfaces/ITimelock.sol
@@ -4,30 +4,30 @@ pragma solidity 0.8.23;
 interface ITimelock {
     event NewAdmin(address indexed newAdmin);
     event NewPendingAdmin(address indexed newPendingAdmin);
-    event NewDelay(uint indexed newDelay);
+    event NewDelay(uint256 indexed newDelay);
     event CancelTransaction(
         bytes32 indexed txHash,
         address indexed target,
-        uint value,
+        uint256 value,
         string signature,
         bytes data,
-        uint eta
+        uint256 eta
     );
     event ExecuteTransaction(
         bytes32 indexed txHash,
         address indexed target,
-        uint value,
+        uint256 value,
         string signature,
         bytes data,
-        uint eta
+        uint256 eta
     );
     event QueueTransaction(
         bytes32 indexed txHash,
         address indexed target,
-        uint value,
+        uint256 value,
         string signature,
         bytes data,
-        uint eta
+        uint256 eta
     );
 
     function delay() external view returns (uint);
@@ -38,25 +38,25 @@ interface ITimelock {
 
     function queueTransaction(
         address target,
-        uint value,
+        uint256 value,
         string calldata signature,
         bytes calldata data,
-        uint eta
+        uint256 eta
     ) external returns (bytes32);
 
     function cancelTransaction(
         address target,
-        uint value,
+        uint256 value,
         string calldata signature,
         bytes calldata data,
-        uint eta
+        uint256 eta
     ) external;
 
     function executeTransaction(
         address target,
-        uint value,
+        uint256 value,
         string calldata signature,
         bytes calldata data,
-        uint eta
+        uint256 eta
     ) external payable returns (bytes memory);
 }
